Recommendation system engine
The file will help you to walk through both pure collaborative filtering and hybrid collaborative filtering.  Generally, the recommendation system engine is made for movies, songs but like amazon or other E-commerce does, I have made this for market basket dataset. 
This was the work as part of my internship.